% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crosswalk_tract_2020}
\alias{crosswalk_tract_2020}
\title{Geographic Crosswalk for 2020 Census Tracts in NYC}
\format{
A data frame with 2,312 rows and 15 variables
\describe{
\item{tract_2020}{full census tract id (2020 geography)}
\item{GEOID10}{10-digit census tract id (2020 geography)}
\item{tract}{census tract (2020 geography)}
\item{tract_name}{census tract name (2020 geography)}
\item{borough }{borough}
\item{neighborhood}{neighborhood}
\item{nCode}{unique neighborhood code}
\item{congressional}{congressional district}
\item{stateSenate}{state senate district}
\item{council}{city council district}
\item{assembly}{assembly district}
\item{school}{school district}
\item{commBoard}{community board district}
\item{zip}{zip code}
\item{modzcta}{modified zip code tabulation area}
}
}
\usage{
crosswalk_tract_2020
}
\description{
A dataset containing a crosswalk file listing the various geographies each
2020 census tract is assigned to in New York City
}
\keyword{datasets}
