% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demos_data}
\alias{demos_data}
\title{American Community Survey Data for NYC Assembly Districts}
\format{
A large list containing 10 elements representing different geographic levels. Each element contains a data frame with 162 demographic variables.
}
\usage{
demos_data
}
\description{
A dataset containing demographic data from the American Community Survey.
 Includes information on: total population, total housing units (rented vs. owned),
 median age, median income, sex, poverty, SNAP recipients, age, race, household
 type, marital status, education level, and language. All variables are from the
 2019 ACS 5-yr estimates except language data, which is only available below the
 county level through the 2015 ACS.
}
\details{
Data available for the following NYC geographies:
 - assembly districts (assembly)
 - boroughs (borough)
 - community boards (commBoard)
 - congressional districts (congressional)
 - city council districts (council)
 - zip code (modzcta)
 - neighborhood (nCode)
 - school district (school)
 - state senate district (stateSenate)
 - census tracts (tract_2010)
}
\keyword{datasets}
