% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demos_nCode}
\alias{demos_nCode}
\title{American Community Survey Data for NYC Neighborhoods}
\format{
A data frame with 252 rows and 164 variables
}
\usage{
demos_nCode
}
\description{
A dataset containing demographic data from the American Community Survey.
Includes information on: total population, total housing units (rented vs. owned),
median age, median income, sex, poverty, SNAP recipients, age, race, household
type, marital status, education level, and language. All variables are from the
2019 ACS 5-yr estimates except language data, which is only available below the
county level through the 2015 ACS.
}
\keyword{datasets}
