% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_assembly}
\alias{geo_assembly}
\title{Geographic Shapefile for NYC}
\format{
A sf data frame with 65 rows and 2 variables
\describe{
\item{GEO_ID}{assembly district}
\item{geometry}{multipolygon geometry}
}
}
\usage{
geo_assembly
}
\description{
A dataset containing a simple features (sf) data frame of NYC assembly districts
}
\keyword{datasets}
