% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_borough}
\alias{geo_borough}
\title{Geographic Shapefile for NYC}
\format{
A sf data frame with 5 rows and 2 variables
\describe{
\item{GEO_ID}{borough}
\item{geometry}{multipolygon geometry}
}
}
\usage{
geo_borough
}
\description{
A dataset containing a simple features (sf) data frame of NYC boroughs
}
\keyword{datasets}
