% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_commBoard}
\alias{geo_commBoard}
\title{Geographic Shapefile for NYC}
\format{
A sf data frame with 71 rows and 2 variables
\describe{
  \item{GEO_ID}{community board district}
  \item{geometry}{multipolygon geometry}
}
}
\usage{
geo_commBoard
}
\description{
A dataset containing a simple features (sf) data frame of NYC community board districts
}
\keyword{datasets}
