% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_congressional}
\alias{geo_congressional}
\title{Geographic Shapefile for NYC}
\format{
A sf data frame with 13 rows and 2 variables
\describe{
\item{GEO_ID}{congressional district}
\item{geometry}{multipolygon geometry}
}
}
\usage{
geo_congressional
}
\description{
A dataset containing a simple features (sf) data frame of NYC congressional districts
}
\keyword{datasets}
