% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_council}
\alias{geo_council}
\title{Geographic Shapefile for NYC}
\format{
A sf data frame with 51 rows and 2 variables
\describe{
\item{GEO_ID}{council district}
\item{geometry}{multipolygon geometry}
}
}
\usage{
geo_council
}
\description{
A dataset containing a simple features (sf) data frame of NYC council districts
}
\keyword{datasets}
