% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_data}
\alias{geo_data}
\title{Geographic Shapefiles for NYC}
\format{
A large list containing 11 elements representing different geographic levels. Each element contains a sf data frame with 2 variables:
\describe{
  \item{GEO_ID}{geographic identifier}
  \item{geometry}{multipolygon geometry}
}
}
\usage{
geo_data
}
\description{
A list containing a simple features (sf) data frame for each of the following NYC geographies:
 - assembly districts (assembly)
 - boroughs (borough)
 - community boards (commBoard)
 - congressional districts (congressional)
 - city council districts (council)
 - zip code (modzcta)
 - neighborhood (nCode)
 - school district (school)
 - state senate district (stateSenate)
 - census tracts (tract_2010 and tract_2020)
}
\keyword{datasets}
