% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_modzcta}
\alias{geo_modzcta}
\title{Geographic Shapefile for NYC}
\format{
A sf data frame with 178 rows and 2 variables
\describe{
\item{GEO_ID}{modzcta (modified zip code tabulation area)}
\item{geometry}{multipolygon geometry}
}
}
\usage{
geo_modzcta
}
\description{
A dataset containing a simple features (sf) data frame of NYC zip codes (modified zip code
tabulation areas)
}
\keyword{datasets}
