% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_nCode}
\alias{geo_nCode}
\title{Geographic Shapefile for NYC}
\format{
A sf data frame with 253 rows and 2 variables
\describe{
\item{GEO_ID}{neighborhood code (nCode)}
\item{geometry}{multipolygon geometry}
}
}
\usage{
geo_nCode
}
\description{
A dataset containing a simple features (sf) data frame of NYC neighborhoods
}
\keyword{datasets}
