% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_school}
\alias{geo_school}
\title{Geographic Shapefile for NYC}
\format{
A sf data frame with 32 rows and 2 variables
\describe{
  \item{GEO_ID}{school district}
  \item{geometry}{multipolygon geometry}
}
}
\usage{
geo_school
}
\description{
A dataset containing a simple features (sf) data frame of NYC school districts
}
\keyword{datasets}
