% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_stateSenate}
\alias{geo_stateSenate}
\title{Geographic Shapefile for NYC}
\format{
A sf data frame with 26 rows and 2 variables
\describe{
\item{GEO_ID}{state senate district}
\item{geometry}{multipolygon geometry}
}
}
\usage{
geo_stateSenate
}
\description{
A dataset containing a simple features (sf) data frame of NYC state senate districts
}
\keyword{datasets}
