% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_tract_2020}
\alias{geo_tract_2020}
\title{Geographic Shapefile for NYC}
\format{
A sf data frame with 2197 rows and 2 variables
\describe{
\item{GEO_ID}{census tract (2020 geography)}
\item{geometry}{multipolygon geometry}
}
}
\usage{
geo_tract_2020
}
\description{
A dataset containing a simple features (sf) data frame of NYC census tracts (2020 geography)
}
\keyword{datasets}
