% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_demographics.R
\name{plot_demographics}
\alias{plot_demographics}
\title{Plot demographics}
\usage{
plot_demographics(df)
}
\arguments{
\item{df}{data frame to generate demographic plot from (must contain only one row of data)}
}
\value{
a ggplot object
}
\description{
Generates a stacked bar plot of the proportion of different racial and ethnic groups for a given GEO_ID
}
\examples{
# save demographic data for woodside
woodside <- nyccensus::demos_data[["nCode"]] \%>\%
  dplyr::filter(GEO_ID == "Q64")

# generate plot
plot_demographics(woodside)
}
