% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rr_congressional}
\alias{rr_congressional}
\title{Daily Response Rate Data for the 2020 Census in New York City}
\format{
A data frame with 2,223 rows and 6 variables:
\describe{
  \item{RESP_DATE}{response date (format: YYYY-MM-DD)}
  \item{GEO_ID}{congressional district}
  \item{CRRALL}{cumulative response rate}
  \item{CRRINT}{cumulative response rate (internet-only)}
  \item{DRRALL}{daily response rate}
  \item{DRRINT}{daily response rate (internet-only)}
}
}
\usage{
rr_congressional
}
\description{
A dataset containing daily response rate data (March 20 to October 17) reported
 by the US Census Bureau for the 2020 Census for NYC congressional districts.
 The variables are as follows:
}
\keyword{datasets}
