% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rr_data}
\alias{rr_data}
\title{Daily Response Rate Data for the 2020 Census in New York City}
\format{
A large list containing 10 elements representing different geographic levels. Each element contains the following variables:
\describe{
  \item{RESP_DATE}{response date (format: YYYY-MM-DD)}
  \item{GEO_ID}{geographic identifier}
  \item{CRRALL}{cumulative response rate}
  \item{CRRINT}{cumulative response rate (internet-only)}
  \item{DRRALL}{daily response rate}
  \item{DRRINT}{daily response rate (internet-only)}
}
}
\usage{
rr_data
}
\description{
A dataset containing daily response rate data (March 20 to October 17) reported
 by the US Census Bureau for the 2020 Census for the following NYC geographies:
 - assembly districts (assembly)
 - boroughs (borough)
 - community boards (commBoard)
 - congressional districts (congressional)
 - city council districts (council)
 - zip code (modzcta)
 - neighborhood (nCode)
 - school district (school)
 - state senate district (stateSenate)
 - census tracts (tract_2020)
}
\keyword{datasets}
