% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rr_nCode}
\alias{rr_nCode}
\title{Daily Response Rate Data for the 2020 Census in New York City}
\format{
A data frame with 42,750 rows and 8 variables:
\describe{
\item{RESP_DATE}{response date (format: YYYY-MM-DD)}
\item{GEO_ID}{neighborhood code (nCode)}
\item{BORO}{borough}
\item{NEIGHBORHOOD}{neighborhood}
\item{CRRALL}{cumulative response rate}
\item{CRRINT}{cumulative response rate (internet-only)}
\item{DRRALL}{daily response rate}
\item{DRRINT}{daily response rate (internet-only)}
}
}
\usage{
rr_nCode
}
\description{
A dataset containing daily response rate data (March 20 to October 17) reported
by the US Census Bureau for the 2020 Census for NYC neighborhoods.
The variables are as follows:
}
\keyword{datasets}
