% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_census.R
\name{theme_census}
\alias{theme_census}
\title{Census Theme}
\usage{
theme_census(...)
}
\value{
a theme function
}
\description{
Applies standard plotting theming to ggplot objects
}
\examples{
ggplot2::ggplot(iris, ggplot2::aes(x = Sepal.Length)) +
  ggplot2::geom_histogram() +
  theme_census()
}
